refactor: move functions out of main file
